Alpha Release 1.0.2
Fixed all errors that I ran into with the Unix Version and it is now released and sorta functional although it sucks. But it is just the concept being thrown together. Sadly windows support has been dropped due to my lack of knowledge on how to develop anything for windows and the fact that it is just simply very slow. I was able to fix errors with Unix in 1/3 of the time. And still never got windows up and running. By the time Safron reaches a Stable release is when I will start on the Gui.  [Stable.Beta.Alpha] is how my releases will look. so once I am at 1.0.0, I will start on the Graphical version which will be a seperate repo. I will not be promise a release time on Windows GUI but probably On the 5th stable release 

Signed-off-by: Asher Winstead <git@asherwinstead.dev>